sentry in raftcat
