+++ raftcat/Cargo.toml
@@ -17,3 +17,5 @@ actix-web = "0.7.13"
 env_logger = "0.5.13"
 actix = "0.7.6"
 log = "0.4.6"
+sentry = "0.12.0"
+sentry-actix = "0.12.0"
