+++ raftcat/src/kube.rs
@@ -1,0 +1,75 @@
+use kubernetes::client::APIClient;
+use std::collections::HashMap;
+use shipcat_definitions::{Crd, CrdList, Manifest};
+
+use failure::{Error, Fail};
+pub type Result<T> = std::result::Result<T, Error>;
+
+static GROUPNAME: &str = "babylontech.co.uk";
+static SHIPCATRESOURCE: &str = "shipcatmanifests";
+
+
+// Request builders
+fn make_all_crd_entry_req(resource: &str, group: &str) -> Result<http::Request<Vec<u8>>> {
+    let urlstr = format!("/apis/{group}/v1/namespaces/dev/{resource}?",
+        group = group, resource = resource);
+    let urlstr = url::form_urlencoded::Serializer::new(urlstr).finish();
+    let mut req = http::Request::get(urlstr);
+    req.body(vec![]).map_err(Error::from)
+}
+fn make_crd_entry_req(resource: &str, group: &str, name: &str) -> Result<http::Request<Vec<u8>>> {
+    // TODO: namespace from evar
+    let urlstr = format!("/apis/{group}/v1/namespaces/dev/{resource}/{name}?",
+        group = group, resource = resource, name = name);
+    let urlstr = url::form_urlencoded::Serializer::new(urlstr).finish();
+    let mut req = http::Request::get(urlstr);
+    req.body(vec![]).map_err(Error::from)
+}
+/*fn watch_crd_entry_after(resource: &str, group: &str, name: &str, rver: u32) -> Result<http::Request<Vec<u8>>> {
+    let urlstr = format!("/apis/{group}/v1/namespaces/dev/{resource}/{name}?",
+        group = group, resource = resource, name = name);
+    let mut qp = url::form_urlencoded::Serializer::new(urlstr);
+
+    qp.append_pair("timeoutSeconds", "30");
+    qp.append_pair("watch", "true");
+
+    // last version to watch after
+    //qp.append_pair("resourceVersion", &rver.to_string());
+
+    let urlstr = qp.finish();
+    let mut req = http::Request::get(urlstr);
+    req.body(vec![]).map_err(Error::from)
+}*/
+
+
+// program interface - request consumers
+pub type ManifestMap = HashMap<String, Crd<Manifest>>;
+
+pub fn get_shipcat_manifests(client: &APIClient) -> Result<ManifestMap> {
+    let req = make_all_crd_entry_req(SHIPCATRESOURCE, GROUPNAME)?;
+    let res = client.request::<CrdList<Manifest>>(req)?;
+    let mut data = HashMap::new();
+    for i in res.items {
+        data.insert(i.spec.name.clone(), i);
+    }
+    let keys = data.keys().cloned().into_iter().collect::<Vec<_>>().join(", ");
+    debug!("Initialized with: {}", keys);
+    Ok(data)
+}
+
+/*this doesn't actually work...
+pub fn watch_shipcat_manifest(client: &APIClient, name: &str, rver: u32) -> Result<Crd<Manifest>> {
+    let req = watch_crd_entry_after(SHIPCATRESOURCE, GROUPNAME, name, rver)
+        .expect("failed to define crd watch request");
+    let res = client.request::<Crd<_>>(req)?;
+    debug!("{}", &res.spec.name);
+    Ok(res)
+}*/
+
+pub fn get_shipcat_manifest(client: &APIClient, name: &str) -> Result<Crd<Manifest>> {
+    let req = make_crd_entry_req(SHIPCATRESOURCE, GROUPNAME, name)?;
+    let res = client.request::<Crd<Manifest>>(req)?;
+    debug!("got {}", &res.spec.name);
+    // TODO: merge with version found in rolling env?
+    Ok(res)
+}
